tests/mock_algorithm/v6-rdf-mock/requirements.txt
vantage6-algorithm-tools
pandas

git+https://github.com/STRONGAYA/v6-tools-general.git@v1.0.2